animation: Move drop target animation code to component.
