+++ VisualPinball.Unity/VisualPinball.Unity/VPT/HitTarget/DropTargetAnimationComponent.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 39ede5c36a2f2a14ca7f17268f5d1e9f
+guid: 1bde8cef51124ae7872f61d7a0d902b8
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
